Refactoring: OCRScanDisplay added helper function for generating dislplay in tests; extracted sources
