+++ tdd_intro/homework/03_bank_ocr/test.cpp
@@ -85,8 +85,6 @@ Example input and output
 ```
 */
 #include <gtest/gtest.h>
-#include <string>
-#include <exception>
 #include "constdata.h"
 #include "display.h"
 
@@ -114,58 +112,16 @@ Example input and output
  * - Not correct digits
 */
 
-bool operator ==(const Digit& left, const Digit& right)
-{
-    bool digitsEqual = true;
-    for (unsigned int i = 0; i < g_linesInDigit; i++)
-    {
-        digitsEqual &= left.lines[i] == right.lines[i];
-    }
-    return digitsEqual;
-}
-
-unsigned int OCRLookupDigit(const Digit& digit)
-{
-    const auto foundDigit = std::find(s_digitsDict.begin(), s_digitsDict.end(), digit);
-    if (foundDigit == s_digitsDict.end())
-    {
-        throw std::exception("Invalid digit.");
-    }
-    unsigned int value = std::distance(s_digitsDict.begin(), foundDigit);
-    return value;
-}
-
-// Positions on display start from 0, from left to right
-void OCRExtractDigitAtPos(const Display& display, unsigned int pos, Digit& target)
+void PopulateDisplay(Display& target, const std::vector<Digit>& digits)
 {
     for (unsigned int i = 0; i < g_linesInDigit; i++)
     {
-        target.lines[i] = display.lines[i].substr(g_digitLen * pos, g_digitLen);
-    }
-}
-
-unsigned int OCRScanDisplay(const Display& display)
-{
-    unsigned int value = 0;
-    try
-    {
-        Digit displayDigit1;
-        OCRExtractDigitAtPos(display, 0, displayDigit1);
-        value = OCRLookupDigit(displayDigit1);
-
-        Digit displayDigit2;
-        OCRExtractDigitAtPos(display, 1, displayDigit2);
-        if (!(displayDigit2 == s_emptyDigit))
+        target.lines[i] = "";
+        for (const auto digit : digits)
         {
-            value *= 10;
-            value += OCRLookupDigit(displayDigit2);
+            target.lines[i] += digit.lines[i];
         }
     }
-    catch (const std::exception& /*ex*/)
-    {
-        //warning "invalid digit"
-    }
-    return value;
 }
 
 TEST(OCRScanDisplay, EmptyDisplay)
@@ -176,34 +132,35 @@ TEST(OCRScanDisplay, EmptyDisplay)
 
 TEST(OCRScanDisplay, SingleDigit1)
 {
-    const Display one = { s_digit1.lines[0], s_digit1.lines[1], s_digit1.lines[2] };
+    Display one;
+    PopulateDisplay(one, {s_digit1});
     EXPECT_EQ(OCRScanDisplay(one), 1);
 }
 
 TEST(OCRScanDisplay, SingleDigit2)
 {
-    const Display two = { s_digit2.lines[0], s_digit2.lines[1], s_digit2.lines[2] };
+    Display two;
+    PopulateDisplay(two, {s_digit2});
     EXPECT_EQ(OCRScanDisplay(two), 2);
 }
 
 TEST(OCRScanDisplay, SingleDigit5)
 {
-    const Display five = { s_digit5.lines[0], s_digit5.lines[1], s_digit5.lines[2] };
+    Display five;
+    PopulateDisplay(five, {s_digit5});
     EXPECT_EQ(OCRScanDisplay(five), 5);
 }
 
 TEST(OCRScanDisplay, TwoDigits10)
 {
-    const Display ten = { s_digit1.lines[0] + s_digit0.lines[0],
-                          s_digit1.lines[1] + s_digit0.lines[1],
-                          s_digit1.lines[2] + s_digit0.lines[2] };
+    Display ten;
+    PopulateDisplay(ten, {s_digit1, s_digit0});
     EXPECT_EQ(OCRScanDisplay(ten), 10);
 }
 
 TEST(OCRScanDisplay, TwoDigits01)
 {
-    const Display one = { s_digit0.lines[0] + s_digit1.lines[0],
-                          s_digit0.lines[1] + s_digit1.lines[1],
-                          s_digit0.lines[2] + s_digit1.lines[2] };
+    Display one;
+    PopulateDisplay(one, {s_digit0, s_digit1});
     EXPECT_EQ(OCRScanDisplay(one), 1);
 }
